Wrapped std::time::Duration
