+++ templates/index.html
@@ -16,7 +16,7 @@
 	    </div>
 	    <div class="tile-content">
 	      <h3 class="tile-title text-centered ">{{ activity.session.start_time }}</h3>
-	      <p class="tile-subtitle text-centered">{{ activity.session.duration_active.to_string() }}</p>
+	      <p class="tile-subtitle text-centered">{{ activity.session.duration_active }}</p>
 	    </div>
 	    <div class="tile-action">
 	      <img src="{{ url._static }}/img/activity/{{ activity.id }}.png" class="img-fit-contain activity-image" height="150">
